+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- [fixed] when an all watched eng/ger season was opened the visual didn't change
 - [feature] added feature to compare the local series list against an external list to see which series we are missing
   - then displaying the missing series in a modal
   - under the add tab